added quantity chenge on deleting order

+++ server/server/controllers/orderController.js
@@ -175,8 +175,7 @@ const getOrderDetails = async (req, res, next) => {
     }
 };
 
-// Update an existing order
-// Update an existing order
+
 const updateOrder = async (req, res, next) => {
     try {
         const orderId = req.params.id;
@@ -211,11 +210,25 @@ const updateOrder = async (req, res, next) => {
 };
 
 
-// Delete an existing order
+
+
 const deleteOrder = async (req, res, next) => {
     try {
         const orderId = req.params.id;
 
+        const order = await Order.findById(orderId);
+
+        if (!order) {
+            res.status(404);
+            return next(new Error('Order not found'));
+        }
+
+        // Restore the quantity of ticket items to the available tickets
+        for (const item of order.ticketItems) {
+            await Ticket.findByIdAndUpdate(item.ticketId, { $inc: { availableTickets: item.quantity } });
+        }
+
+        // Delete the order
         const deletedOrder = await Order.findByIdAndDelete(orderId);
 
         if (!deletedOrder) {
@@ -229,6 +242,7 @@ const deleteOrder = async (req, res, next) => {
         return next(new Error('Failed to delete order'));
     }
 };
+
 
 module.exports = { createOrder, getUserOrders, getOrderDetails, updateOrder, deleteOrder };
